+++ packages/store/src/execution/symbols.ts
@@ -1,0 +1,7 @@
+/*
+ * Execution strategy interface
+ */
+export interface NgxsExecutionStrategy {
+  enter<T>(func: () => T): T;
+  leave<T>(func: () => T): T;
+}